+++ CHANGELOG.markdown
@@ -1,9 +1,10 @@
 5.3.8 [unreleased]
 ------------------
 
-* Add `Generic1`-based functions for many classes:
+* Add `Generic1`-based functions for many classes, useful for writing instances:
   - `Data.Functor.Alt.(<!>)` -> `Data.Functor.Alt.galt`
   - `Data.Functor.Apply.{liftF2,liftF3}` -> `Data.Functor.Apply.{gliftF2,gliftF3}`
+  - `Data.Functor.Bind.(>>-)` -> `Data.Functor.Bind.gbind`
 
 5.3.7 [2022.01.09]
 ------------------